New file structure and refreshed layout
