+++ README.md
@@ -1,38 +1,37 @@
 Sinatra Realtime Chat
 ---------------------
 
-A little chat app using sinatra streaming!
+# WORK IN PROGRESS (3 Aug 2014), KEEP CALM AND DRINK ALE
 
-You can try it online [Sinatra realtime app](http://sinatra-realtime-chat.herokuapp.com)
+Goals
+-----
 
-What can you expect from this repo
-----------------------------------
+* Decent websocket chat with AngularJS frontend
 
-So this is just a pet project I started to learn about:
-
-* sinatra: Looking forward to building simpler and faster apps
-* Caching: Investigating about how http and browsers handle caching and be able to improve perfomance this way. I found this article about [different ways of caching](http://betterexplained.com/articles/how-to-optimize-your-site-with-http-caching/)
-* SSE (Server Sent Events): This is why I started with sinatra, I'm trying to build more realtime interaction apps with ruby and found out that It's already built in with this framework. Basics about SSE can be found in this article: [HTML5Rocks SSE Basics](http://www.html5rocks.com/en/tutorials/eventsource/basics/)
-* Integration testing: I've been looking for a project to make integration testing with Capybara.
-
-Dependencies
+Bootstrapping
 ------------
 
-Main:
+Easy things first: start your PostgreSQL server and create database `chatdb`.
 
-* sinatra
-* data_mapper
-* thin (seems like Webrick just is not able to handle connections)
-* pg (Postgress db access)
+Then clone this repo and run the following commands:
 
-Testing:
+```
+$ bundle install
+$ npm install -g gulp bower
+$ bower install
+$ rackup
+```
 
-* capybara
-* poltergeist
-* rake
-* xpath
+In a separate session run:
 
-Author
+```
+$ gulp
+```
+
+App will be available under link `http://0.0.0.0:9292`
+
+Author(s)
 ------
 
 [Alvaro F. Lara](http://alvarofernandolara.com.ar)
+[Raf Szalanski](http://szalansky.com)
