+++ views/chat.erb
@@ -7,7 +7,6 @@
     <meta name="description" content="">
     <meta name="author" content="">
 
-    <link href="/css/bootstrap.min.css" rel="stylesheet">
     <style type="text/css">
       body {
         padding-top: 60px;
@@ -17,7 +16,7 @@
         padding: 9px 0;
       }
     </style>
-    <link href="/css/bootstrap-responsive.min.css" rel="stylesheet">
+    <link href="/bower/foundation/css/foundation.css" rel="stylesheet">
     <link href="/stylesheets/application.css" rel="stylesheet">
 
     <!-- HTML5 shim, for IE6-8 support of HTML5 elements -->
@@ -27,46 +26,34 @@
   </head>
 
   <body ng-app="ChatApp">
-    <div class="navbar navbar-inverse navbar-fixed-top">
-      <div class="navbar-inner">
-        <div class="container-fluid">
-          <a class="brand" href="/">Rafchat</a>
+    <nav class="top-bar"></nav>
+    <div class="row full-width">
+      <div class="large-3 medium-3 columns users" ng-controller="UserController">
+        <ul class="side-nav">
+          <li ng-repeat="user in users"><a href="#">{{user.username}}</a></li>
+        </ul>
+      </div>
+
+      <div class="large-9 medium-9 columns chat" ng-controller="ChatController">
+        <div class="messages">
+          <div class="msg" ng-repeat="msg in messages">
+            <div class="owner">{{msg.owner}}</div>
+            <div class="content">
+              <div class="body">
+                {{msg.body}}
+              </div>
+              <div class="timestamp">
+                {{msg.created_at}}
+              </div>
+            </div>
+          </div>
+        </div>
+
+        <div class="message-input">
+          <textarea ng-model="message"></textarea>
+          <button ng-click="sendMessage()">Send</button>
         </div>
       </div>
-    </div>
-
-    <div class="container-fluid">
-      <div class="row-fluid">
-        <div class="span3" ng-controller="UserController">
-          <div class="well sidebar-nav general_dimensions">
-            <ul class="nav nav-list" id="user_list">
-              <li class="nav-header"><a href="/logout"><strong>Logout</strong></a></li>
-              <li class="nav-header">Online users</li>
-              <li ng-repeat="user in users">
-                <a href="#" id="{{user.username}}">{{user.username}}</a>
-              </li>
-            </ul>
-          </div>
-        </div>
-        <div class="span9 general_dimensions" id="chat" ng-controller="ChatController">
-          <div class="row-fluid chat-window" scroll-glue>
-            <p class="message" ng-repeat="msg in messages track by $index">
-              <strong>{{msg.owner}}</strong>: <span title="{{msg.created_at | dateify | date:'hh:mm a'}}">{{msg.body}}</span>
-            </p>
-          </div>
-
-          <form class="row-fluid" id="controls" ng-submit="sendMessage()">
-            <input autofocus type="text" id="message_input" name="message" placeholder="Say it here." class="span10 pull-left" ng-model="message">
-            <input type="submit" class="btn span2 pull-right" value="Say it!">
-          </form>
-        </div><!--/span-->
-      </div><!--/row-->
-      <hr>
-
-      <footer>
-        <p>&copy; Who's Company 2012</p>
-      </footer>
-
     </div>
     <script type="text/javascript" src="/bower/jquery/dist/jquery.js"></script>
     <script type="text/javascript" src="/bower/angular/angular.js"></script>
